+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/event/CaseEventReceiverSpringTest.java
@@ -1,21 +1,19 @@
-package uk.gov.ons.ctp.integration.rhsvc.event.impl;
+package uk.gov.ons.ctp.integration.rhsvc.event;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.parallel.ResourceAccessMode.READ_WRITE;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.wildfly.common.Assert.assertTrue;
 
 import com.google.cloud.spring.pubsub.core.PubSubTemplate;
 import com.google.cloud.spring.pubsub.integration.inbound.PubSubInboundChannelAdapter;
 import java.text.SimpleDateFormat;
 import java.util.HashMap;
 import java.util.Locale;
-import java.util.Set;
 import java.util.TimeZone;
-import lombok.SneakyThrows;
+import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -32,10 +30,9 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.ons.ctp.common.FixtureHelper;
-import uk.gov.ons.ctp.common.event.EventTopic;
-import uk.gov.ons.ctp.common.event.model.UacEvent;
+import uk.gov.ons.ctp.common.event.model.CaseEvent;
+import uk.gov.ons.ctp.common.event.model.Header;
 import uk.gov.ons.ctp.common.utility.ParallelTestLocks;
-import uk.gov.ons.ctp.integration.rhsvc.RespondentHomeFixture;
 import uk.gov.ons.ctp.integration.rhsvc.config.AppConfig;
 import uk.gov.ons.ctp.integration.rhsvc.repository.CaseRepository;
 import uk.gov.ons.ctp.integration.rhsvc.repository.CollectionExerciseRepository;
@@ -49,16 +46,15 @@ import uk.gov.ons.ctp.integration.rhsvc.repository.UacRepository;
 @ExtendWith(SpringExtension.class)
 @ActiveProfiles("mocked-connection-factory")
 @ResourceLock(value = ParallelTestLocks.SPRING_TEST, mode = READ_WRITE)
-public class UacEventReceiverImplSpringTest {
+public class CaseEventReceiverSpringTest {
 
-  @Autowired private PubSubInboundChannelAdapter uacEventInbound;
-  @Autowired private AppConfig appConfig;
-  @Autowired private UACEventReceiver receiver;
-  @MockBean private PubSubTemplate pubSubTemplate;
+  @Autowired private CaseEventReceiver receiver;
+  @Autowired private PubSubInboundChannelAdapter caseEventInbound;
   @MockBean private SurveyRepository respondentSurveyRepo;
   @MockBean private CollectionExerciseRepository respondentCollExRepo;
   @MockBean private CaseRepository respondentCaseRepo;
   @MockBean private UacRepository respondentUacRepo;
+  @MockBean private PubSubTemplate pubSubTemplate;
   @MockBean private EventFilter eventFilter;
 
   @BeforeEach
@@ -66,78 +62,78 @@ public class UacEventReceiverImplSpringTest {
     Mockito.reset(receiver);
   }
 
-  @SneakyThrows
-  private void UacEventFlow(EventTopic topic) {
-    UacEvent uacEvent = FixtureHelper.loadPackageFixtures(UacEvent[].class).get(0);
-    uacEvent.getPayload().getUacUpdate().setQid(RespondentHomeFixture.A_QID);
+  /** Test the receiver flow */
+  @Test
+  public void caseEventFlowTest() throws Exception {
+    Header header = new Header();
+    header.setMessageId(UUID.fromString("c45de4dc-3c3b-11e9-b210-d663bd873d93"));
+    CaseEvent caseEvent = FixtureHelper.loadPackageFixtures(CaseEvent[].class).get(0);
+    caseEvent.setHeader(header);
 
     // Construct message
-    Message<UacEvent> message = new GenericMessage<>(uacEvent, new HashMap<>());
+    Message<CaseEvent> message = new GenericMessage<>(caseEvent, new HashMap<>());
 
     when(eventFilter.isValidEvent(any(), any(), any(), any())).thenReturn(true);
 
     // Send message to container
-    uacEventInbound.getOutputChannel().send(message);
+    caseEventInbound.getOutputChannel().send(message);
 
     // Capture and check Service Activator argument
-    ArgumentCaptor<UacEvent> captur = ArgumentCaptor.forClass(UacEvent.class);
-    verify(receiver).acceptUACEvent(captur.capture());
-    assertTrue(captur.getValue().getPayload().equals(uacEvent.getPayload()));
-    verify(respondentUacRepo).writeUAC(any());
+    ArgumentCaptor<CaseEvent> captur = ArgumentCaptor.forClass(CaseEvent.class);
+    verify(receiver).acceptCaseEvent(captur.capture());
+    assertEquals(captur.getValue().getPayload(), caseEvent.getPayload());
+    verify(respondentCaseRepo, times(1)).writeCaseUpdate(caseEvent.getPayload().getCaseUpdate());
   }
 
-  /** Test the receiver flow for UAC updated */
+  /** Test the receiver flow for SIS case event */
   @Test
-  public void uacUpdatedEventFlow() {
-    UacEventFlow(EventTopic.UAC_UPDATE);
-  }
-
-  @Test
-  public void shouldFilterUacEventWithContinuationFormQid() throws Exception {
-
-    appConfig.getQueueConfig().setQidFilterPrefixes(Set.of("12"));
-    UacEvent uacEvent = FixtureHelper.loadPackageFixtures(UacEvent[].class).get(0);
-    uacEvent.getPayload().getUacUpdate().setQid(RespondentHomeFixture.QID_12);
+  public void caseEventSISFlowTest() throws Exception {
+    Header header = new Header();
+    header.setMessageId(UUID.fromString("c45de4dc-3c3b-11e9-b210-d663bd873d93"));
+    CaseEvent caseEvent = FixtureHelper.loadPackageFixtures(CaseEvent[].class).get(0);
+    caseEvent.setHeader(header);
 
     // Construct message
-    Message<UacEvent> message = new GenericMessage<>(uacEvent, new HashMap<>());
+    Message<CaseEvent> message = new GenericMessage<>(caseEvent, new HashMap<>());
 
-    when(eventFilter.isValidEvent(any(), any(), any(), any())).thenReturn(true);
+    when(eventFilter.isValidEvent(any(), any(), any(), any())).thenReturn(false);
 
     // Send message to container
-    uacEventInbound.getOutputChannel().send(message);
+    caseEventInbound.getOutputChannel().send(message);
 
     // Capture and check Service Activator argument
-    ArgumentCaptor<UacEvent> captur = ArgumentCaptor.forClass(UacEvent.class);
-    verify(receiver).acceptUACEvent(captur.capture());
-    assertTrue(captur.getValue().getPayload().equals(uacEvent.getPayload()));
-    verify(respondentUacRepo, never()).writeUAC(any());
-    verify(respondentUacRepo, never()).writeUAC(any());
+    ArgumentCaptor<CaseEvent> captur = ArgumentCaptor.forClass(CaseEvent.class);
+    verify(receiver).acceptCaseEvent(captur.capture());
+    assertEquals(captur.getValue().getPayload(), caseEvent.getPayload());
+    verify(respondentCaseRepo, times(0)).writeCaseUpdate(caseEvent.getPayload().getCaseUpdate());
   }
 
   @Test
-  public void UacEventReceivedWithoutMillisecondsTest() throws Exception {
+  public void caseCaseReceivedWithoutMillisTest() throws Exception {
 
-    // Create a UAC with a timestamp. Note that the milliseconds are not specified
-    UacEvent uacEvent = FixtureHelper.loadPackageFixtures(UacEvent[].class).get(0);
-    uacEvent.getPayload().getUacUpdate().setQid(RespondentHomeFixture.A_QID);
+    // Create a case with a timestamp. Note that the milliseconds are not specified
+    CaseEvent caseEvent = FixtureHelper.loadPackageFixtures(CaseEvent[].class).get(0);
+
+    Header header = new Header();
     SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.ENGLISH);
     sdf.setTimeZone(TimeZone.getTimeZone("GMT"));
-    uacEvent.getHeader().setDateTime(sdf.parse("2011-08-12T20:17:46Z"));
+    header.setDateTime(sdf.parse("2011-08-12T20:17:46Z"));
+    header.setMessageId(UUID.fromString("c45de4dc-3c3b-11e9-b210-d663bd873d93"));
+    caseEvent.setHeader(header);
 
     // Construct message
-    Message<UacEvent> message = new GenericMessage<>(uacEvent, new HashMap<>());
+    Message<CaseEvent> message = new GenericMessage<>(caseEvent, new HashMap<>());
 
     when(eventFilter.isValidEvent(any(), any(), any(), any())).thenReturn(true);
+
     // Send message to container
-    uacEventInbound.getOutputChannel().send(message);
+    caseEventInbound.getOutputChannel().send(message);
 
     // Capture and check Service Activator argument
-    ArgumentCaptor<UacEvent> captur = ArgumentCaptor.forClass(UacEvent.class);
-    verify(receiver).acceptUACEvent(captur.capture());
+    ArgumentCaptor<CaseEvent> captur = ArgumentCaptor.forClass(CaseEvent.class);
+    verify(receiver).acceptCaseEvent(captur.capture());
     assertEquals(sdf.parse("2011-08-12T20:17:46Z"), captur.getValue().getHeader().getDateTime());
-    assertEquals(uacEvent.getHeader(), captur.getValue().getHeader());
-    assertTrue(captur.getValue().getPayload().equals(uacEvent.getPayload()));
-    verify(respondentUacRepo).writeUAC(any());
+    assertEquals(caseEvent.getHeader(), captur.getValue().getHeader());
+    assertEquals(caseEvent.getPayload(), captur.getValue().getPayload());
   }
 }
