+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/event/CollectionExerciseEventReceiverUnit_test.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhsvc.event.impl;
+package uk.gov.ons.ctp.integration.rhsvc.event;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.verify;
