+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/event/MessageSpringConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhsvc.event.impl;
+package uk.gov.ons.ctp.integration.rhsvc.event;
 
 import org.mockito.Mockito;
 import org.springframework.context.annotation.Bean;
