+++ src/test/resources/uk/gov/ons/ctp/integration/rhsvc/event/PackageFixture.CaseEvent.json
